+++ CHANGELOG.md
@@ -8,6 +8,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+
+- responses: replication_state and done_percentage to volume
+
 ## [1.24.0] - 2024-12-17
 
 ### Added